pyproject.toml
[tool.black]
line-length = 115
target-version = ['py313']

[tool.pdm.build]
[tool.pdm.dev-dependencies]
lint = [
    "black>=25.1.0",
    "pyright>=1.1.396",
]
tools = [
    "pre-commit>=4.1.0,<5",
]

[project]
name = "discord-message-scheduler"
version = "1.2.3"
description = "A simple Discord Bot that can send scheduled messages."
authors = [
    {name = "Taku", email = "noemail@example.com"},
]
dependencies = [
    "aiosqlite==0.21",
    "arrow==1.3",  # Doesn't officially support Python 3.13 yet
    "dateparser==1.2.1",
    "discord.py==2.5.2",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.0",
    "packaging>=24.2",
    "python-dateutil~=2.9",  # Doesn't officially support Python 3.13 yet
    "python-dotenv<1.1,>=1.0.1",
    "rich~=13.9.4",
    "sqlalchemy>=2.0.0",
    "tomli~=2.2.1",
    "uvicorn>=0.32.0",
]
requires-python = ">=3.13"
license = {text = "GPL-3.0-only"}

[project.optional-dependencies]
speed = [
    "discord.py[speed]==2.5.2",
    "uvloop>=0.21.0; platform_system != \"Windows\"",
]

[tool.pyright]
exclude = ["**/node_modules",
    "**/__pycache__",
    "**/venv",
    "**/.venv",
]

reportMissingImports = true

pythonVersion = "3.13"
pythonPlatform = "Linux"

typeCheckingMode = "strict"
